requirements.txt
alpaca-py
matplotlib
pandas
python-dotenv
scikit-learn
